pyproject.toml
[tool.poetry]
name = "img-decomposition-api"
version = "0.0.1"
description = "Backend for an app that generates an image, decomposes, classifies, and uncrops it, and searches for product information on identified objects"
authors = ["Christopher Carroll Smith <75859865+chriscarrollsmith@users.noreply.github.com>"]
readme = "README.md"
packages = [
  {include = "img_decomposition"},
  {include = "tests"}
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uuid = "^1.30"
modal-client = "^0.55.4028"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
